Fix bug in get_star_disk_intersection
The conversion to t works on all points on the disk except for one point at the intersection of the disk and the negative x-axis. Thus, if (and only if) this is an intersection point, the quartic equation becomes cubic instead, and the quartic equation solver crashes.

To fix this, the code now checks for this case, and when it occurs uses np.roots to solve the cubic in these cases. Since this occurs rarely, this is done using a for loop. In the other points, where this case does not happen, the code uses the quartic solver as before.